move parallel arith into separate module
